requirements.txt
langchain
langchain-community
langchain-google-genai
langchain-chroma
langchain-text-splitters
chromadb
pypdf
python-dotenv
fastapi
uvicorn